Split out testable functions into helper module
